# public/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CompuWeb API",
        "version": "1.0.0",
        "description": "API REST de **CompuWeb** \u2014 Sistema de inventario y ventas para MyPEs.\n\n## Autenticaci\u00f3n\n\nLos endpoints protegidos requieren un token Bearer (Laravel Passport).\n\n1. Obt\u00e9n tu token en `POST /api/auth/login`\n2. \u00dasalo en cada request: `Authorization: Bearer {token}`\n\n## Niveles de acceso\n\n| Nivel | Descripci\u00f3n |\n|-------|-------------|\n| P\u00fablico | Sin token \u2014 cat\u00e1logo de productos de una tienda |\n| Vendedor | Token v\u00e1lido \u2014 registrar y consultar ventas |\n| Admin | Token + role `admin` \u2014 reportes de ventas e inventario |"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/tiendas/{slug}/productos": {
            "get": {
                "operationId": "public.productos",
                "description": "Retorna los productos disponibles (con stock > 0) de una tienda.\nSoporta filtro por categor\u00eda, b\u00fasqueda por texto y ordenamiento.",
                "summary": "Listado de productos",
                "tags": [
                    "Cat\u00e1logo P\u00fablico",
                    "Public"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "minimum": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "from": {
                                            "type": [
                                                "integer",
                                                "null"
                                            ],
                                            "minimum": 1
                                        },
                                        "last_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "minimum": 1
                                        },
                                        "links": {
                                            "type": "array",
                                            "description": "Generated paginator links.",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "url": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "required": [
                                                    "url",
                                                    "label",
                                                    "active"
                                                ]
                                            }
                                        },
                                        "next_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "path": {
                                            "type": [
                                                "string",
                                                "null"
                                            ],
                                            "description": "Base path for paginator generated URLs."
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "description": "Number of items shown per page.",
                                            "minimum": 0
                                        },
                                        "prev_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "to": {
                                            "type": [
                                                "integer",
                                                "null"
                                            ],
                                            "description": "Number of the last item in the slice.",
                                            "minimum": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "description": "Total number of items being paginated.",
                                            "minimum": 0
                                        }
                                    },
                                    "required": [
                                        "current_page",
                                        "data",
                                        "first_page_url",
                                        "from",
                                        "last_page_url",
                                        "last_page",
                                        "links",
                                        "next_page_url",
                                        "path",
                                        "per_page",
                                        "prev_page_url",
                                        "to",
                                        "total"
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/tiendas/{slug}/productos/{id}": {
            "get": {
                "operationId": "public.producto",
                "description": "Retorna la informaci\u00f3n completa de un producto disponible, incluyendo\ncategor\u00eda y atributos adicionales (talla, color, etc.).",
                "summary": "Detalle de un producto",
                "tags": [
                    "Cat\u00e1logo P\u00fablico",
                    "Public"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "nombre": {
                                            "type": "string"
                                        },
                                        "sku": {
                                            "type": "string"
                                        },
                                        "descripcion": {
                                            "type": "string"
                                        },
                                        "precio": {
                                            "type": "string"
                                        },
                                        "stock": {
                                            "type": "string"
                                        },
                                        "unidad_medida": {
                                            "type": "string"
                                        },
                                        "imagen_url": {
                                            "type": "string"
                                        },
                                        "categoria": {
                                            "type": [
                                                "object",
                                                "null"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "nombre": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "nombre"
                                            ]
                                        },
                                        "atributos": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "nombre",
                                        "sku",
                                        "descripcion",
                                        "precio",
                                        "stock",
                                        "unidad_medida",
                                        "imagen_url",
                                        "categoria",
                                        "atributos"
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/reportes/inventario": {
            "get": {
                "operationId": "reporte.inventario",
                "description": "Genera un resumen del estado del inventario: total de productos,\nproductos sin stock, productos con stock bajo y valor total del inventario.\nIncluye una lista de alertas para productos que han alcanzado su stock m\u00ednimo.\nSolo accesible para usuarios con rol `admin`.",
                "summary": "Reporte de inventario",
                "tags": [
                    "Reportes (solo Admin)",
                    "Reporte"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "total_productos": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "sin_stock": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "stock_bajo": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "valor_inventario": {
                                            "type": "string"
                                        },
                                        "alertas": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "total_productos",
                                        "sin_stock",
                                        "stock_bajo",
                                        "valor_inventario",
                                        "alertas"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "Producto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tienda_id": {
                        "type": "integer"
                    },
                    "categoria_id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "sku": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "descripcion": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "imagen": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "precio": {
                        "type": "string"
                    },
                    "precio_costo": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "stock_minimo": {
                        "type": "integer"
                    },
                    "unidad_medida": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "tienda_id",
                    "categoria_id",
                    "nombre",
                    "sku",
                    "descripcion",
                    "imagen",
                    "precio",
                    "precio_costo",
                    "stock",
                    "stock_minimo",
                    "unidad_medida",
                    "estado",
                    "created_at",
                    "updated_at"
                ],
                "title": "Producto"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}